Using function approach in periodic_step_parametric and removed periodic_step_sensitivity
